# Conflicts:
#	src/main/java/com/justLearnIT/dao/SubmitDAO.java
#	src/main/java/com/justLearnIT/dao/SubmitDAOImpl.java
#	src/main/java/com/justLearnIT/model/Problem.java
#	src/main/java/com/justLearnIT/model/Submit.java
#	src/main/java/com/justLearnIT/service/SubmitsServiceImpl.java

+++ src/main/java/com/justLearnIT/controller/LearningTopicController.java
@@ -18,7 +18,7 @@ public class LearningTopicController {
     }
 
     @GetMapping
-    public ResponseEntity<?> getTopic(@PathVariable("topicID") long topicID) {
+    public ResponseEntity<?> getTopic(@PathVariable("topicID") int topicID) {
         return new ResponseEntity(learningTopicService.getTopicById(topicID), HttpStatus.OK);
     }
