+++ src/main/java/com/justLearnIT/dao/TopicDAO.java
@@ -4,7 +4,7 @@ import com.justLearnIT.model.Topic;
 import java.util.List;
 
 public interface TopicDAO {
-    Topic getTopicById(long topicId);
+    Topic getTopicById(int topicId);
     List<Topic> getTopicsByCategoryName(String categoryName);
     void saveOrUpdate(Topic topic);
 }
