+++ src/main/java/com/justLearnIT/service/LearningTopicService.java
@@ -3,5 +3,5 @@ package com.justLearnIT.service;
 import com.justLearnIT.model.Topic;
 
 public interface LearningTopicService {
-    Topic getTopicById(long topicId);
+    Topic getTopicById(int topicId);
 }
